archive_docker_server/requirements.txt
grpcio
grpcio-tools
scylla-driver